+++ packages/contracts-nest/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/logger@0.2.4
+
 ## 3.1.1
 
 ### Patch Changes