+++ packages/domain-user/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 2.5.2
+
+### Patch Changes
+
+- testing lib added
+
+- Updated dependencies []:
+  - @volontariapp/database@1.15.1
+  - @volontariapp/logger@0.2.4
+  - @volontariapp/auth@3.2.4
+  - @volontariapp/contracts@4.0.10
+  - @volontariapp/crypto@0.3.7
+  - @volontariapp/errors@0.5.1
+  - @volontariapp/errors-nest@0.11.1
+
 ## 2.5.1
 
 ### Patch Changes