+++ packages/outbox/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.6.2
+
+### Patch Changes
+
+- testing lib added
+
+- Updated dependencies []:
+  - @volontariapp/messaging@1.1.1
+  - @volontariapp/database@1.15.1
+  - @volontariapp/logger@0.2.4
+  - @volontariapp/config@3.0.1
+  - @volontariapp/errors@0.5.1
+
 ## 0.6.1
 
 ### Patch Changes